DATAUP-515: Fix staging service 500 on deleting export
The staging service cannot currently delete any files or folders created by
the exporter because the staging service runs as user 800 (at least in CI)
and this app runs as root (yuck). However,  both the service and this app
are part of the root group, so adding write permissions to that group should
allow the staging service to delete files. This is a shorter term fix -
the longer term fix would be to use sane user names for both processes,
and possibly have the exporter chown to the staging service.