Fix tests w/ patched SDK
The tests are currently failing because the SDK client in the image is
too old to work with the docker server on even the oldest ubuntu version
in GHA.

This hack uses the patched SDK with newer docker binaries and copies
and alters kb_sdk_actions scripts to run the image as root.

It may be possible in the future to no longer run images as root if we
can figure out why the docker changes currently require it, but that's
another fix.

It should also be noted that the prior sdk image isn't pinnned, which
isn't great. The new image tag doesn't appear to be pinnable either but
that may change in the near future.